Dev 20/08/26

+++ demo/model/ldb.go
@@ -6,15 +6,16 @@ import (
 	"os"
 
 	"github.com/syndtr/goleveldb/leveldb"
-	"github.com/syndtr/goleveldb/leveldb/util"
+	"github.com/syndtr/goleveldb/leveldb/util" //BytesPrefix([]byte)
 )
 
 type DBHandler interface {
 	Create(key string, value string) error
+	/* Read */
+	SelectOne(key string) (Data, error)       //key에 해당하는 Data만 읽어오기
+	SelectAll() ([]Data, error)               //DB 내에 있는 모든 Data 읽어오기
+	SelectList(prefix string) ([]Data, error) //prefix로 시작하는 모든 Data 읽어오기
 	Update(key string, value string) error
-	SelectOne(key string) (Data, error)
-	SelectAll() ([]Data, error)
-	SelectList(prefix string) ([]Data, error)
 	Delete(key string) error
 	Close() error
 }
@@ -29,16 +30,20 @@ type Data struct {
 	Value string `json:"value"`
 }
 
-func NewDBHandler(filepath string) DBHandler {
-	return newLDBHandler(filepath)
+func NewDBHandler(filepath string, env bool) DBHandler {
+	return newLDBHandler(filepath, env)
 }
 
-func newLDBHandler(filepath string) DBHandler {
-	//if filepath already exists, delete all files in the filepath
-	if _, err := os.Stat(filepath); !os.IsNotExist(err) {
-		err := os.RemoveAll(filepath)
-		if err != nil {
-			log.Println(err.Error())
+/* newLDBHandler: DB 열기 */
+func newLDBHandler(filepath string, env bool) DBHandler {
+	//if env == true, 폴더 삭제 권한 있으므로 삭제 가능
+	if env {
+		//폴더가 이미 존재한다면, 폴더를 삭제하기
+		if _, err := os.Stat(filepath); !os.IsNotExist(err) {
+			err := os.RemoveAll(filepath)
+			if err != nil {
+				log.Println(err.Error())
+			}
 		}
 	}
 
@@ -51,11 +56,12 @@ func newLDBHandler(filepath string) DBHandler {
 	return &ldbHandler{db: database}
 }
 
+/* Create: key가 중복인지 확인 후 Put */
 func (l *ldbHandler) Create(key string, value string) error {
-	//search every data in db
+	//DB에 있는 모든 데이터 순회하면서 중복 확인
 	iter := l.db.NewIterator(nil, nil)
 	for iter.Next() {
-		//key is duplicated
+		//key가 중복되는 것이 있다면 error를 return
 		if key == string(iter.Key()) {
 			return errors.New("key is duplicated")
 		}
@@ -66,7 +72,7 @@ func (l *ldbHandler) Create(key string, value string) error {
 		return err
 	}
 
-	//put data
+	//Put
 	err = l.db.Put([]byte(key), []byte(value), nil)
 	if err != nil {
 		return err
@@ -74,25 +80,13 @@ func (l *ldbHandler) Create(key string, value string) error {
 	return nil
 }
 
-func (l *ldbHandler) Update(key string, value string) error {
-	//put만 해도 overwrite 되지만, .ldb에 남아있는 듯 하여 delete->put
-	//batch로 쓰는게 더 빠르다고 해서 batch 사용
-	batch := new(leveldb.Batch)
-	batch.Delete([]byte(key))
-	batch.Put([]byte(key), []byte(value))
-	err := l.db.Write(batch, nil)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
+/* SelectOne: key에 해당하는 Data를 return */
 func (l *ldbHandler) SelectOne(key string) (Data, error) {
 	value, err := l.db.Get([]byte(key), nil)
 
 	var getData Data
 	getData.Key = key
-	getData.Value = string(value)
+	getData.Value = string(value) //value는 byte array
 
 	if err != nil {
 		return getData, err
@@ -100,10 +94,11 @@ func (l *ldbHandler) SelectOne(key string) (Data, error) {
 	return getData, nil
 }
 
+/* SelectAll: DB에 있는 모든 데이터를 List로 만들어서 return */
 func (l *ldbHandler) SelectAll() ([]Data, error) {
 	datas := []Data{}
 
-	//search every data in db
+	//DB에 있는 모든 데이터를 순회하면서 List에 저장
 	iter := l.db.NewIterator(nil, nil)
 	for iter.Next() {
 		var getData Data
@@ -119,6 +114,7 @@ func (l *ldbHandler) SelectAll() ([]Data, error) {
 	return datas, nil
 }
 
+/* SelectList: Prefix를 갖는 모든 데이터를 List로 만들어서 return */
 func (l *ldbHandler) SelectList(prefix string) ([]Data, error) {
 	selectDatas := []Data{}
 
@@ -138,6 +134,21 @@ func (l *ldbHandler) SelectList(prefix string) ([]Data, error) {
 	return selectDatas, nil
 }
 
+/* Update: DB에 있는 기존의 key를 Delete하고 다시 Put */
+func (l *ldbHandler) Update(key string, value string) error {
+	//put만 해도 overwrite 되지만, .ldb에 남아있는 듯 하여 delete->put
+	//batch로 쓰는게 더 빠르다고 해서 batch 사용
+	batch := new(leveldb.Batch)
+	batch.Delete([]byte(key))
+	batch.Put([]byte(key), []byte(value))
+	err := l.db.Write(batch, nil)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+/* Delete: 데이터 삭제 */
 func (l *ldbHandler) Delete(key string) error {
 	err := l.db.Delete([]byte(key), nil)
 	if err != nil {
@@ -146,6 +157,7 @@ func (l *ldbHandler) Delete(key string) error {
 	return nil
 }
 
+/* Close: DB 닫기 */
 func (l *ldbHandler) Close() error {
 	err := l.db.Close()
 	if err != nil {
